feat: Move current collateral functionality into decorators
This commit renames the read_keyring decorator to open_keyring and
wraps it in a function body, so that the developer is able
to specify whether the keyring needs to be saved afterwards.

It also adds a validate_label decorator which checks whether the given label
exists. Passing it the exists=False property checks the inverse and is useful
for methods such as add_key.