fix(keyring): Fall back to empty dict if JSON has another type
In cases where the keyring file contained valid JSON, but not in a format that yields `dict` (such as a file affected by the bug described in issue #15), subsequent commands would fail since the object passed was of the wrong type.
This commit adds a check that `keyring` is an instance of `dict`, and overrides it with a dict otherwise.