fix(cli): Return valid keyring if required argument count is not met
In cases where the required argument count was not met, the decorator would return None instead of the keyring.

This commit changes the behavior of the decorator so it returns the keyring if it is available.

It also makes it necessary to decorate a method with open_keyring before required_argument_count to better cover malformed input.

Fixes #15